Remove the Item generic parameter... for now.
Note that while it would be possible--and very useful for buffers of
structs in e.g. system call interfaces (like struct linux_dirent64), and
initialize them lazily in safe code--allowing a generic item type, would
also allow that type to be MaybeUninit<T> in the current type system,
causing endless cycles.

If the compiler would ever support things like specialization or
negative trait impls, this would be achievable with the current
interface, or the Initialize trait could be further reduced to only
MaybeUninit types, with a conversion function for already initialized
types.