Add with_current_vector_unfilled_zeroed.
The main reason for this method, is to be able to quickly initialize the
uninitialized part of the current vector, to later be able to pass it to
less sophisticated I/O code that assume initializedness. A good thing
with the stateful buffer however, is that only uninitialized bytes are
zeroed, so nothing is zeroed twice.